services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication and authorization service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "passlib[bcrypt]>=1.7",
    "pydantic-settings>=2.7",
    "email-validator>=2.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth_service"]
